[DZ-7] Update : Add url domain 'auth'. Separate UserController to AuthController and UserUtilController
