+++ src/main/java/com/dungzi/backend/global/config/security/jwt/JwtAuthenticationFilter.java
@@ -20,7 +20,7 @@ import java.io.IOException;
 @RequiredArgsConstructor
 //@WebFilter(urlPatterns = {"/api/v1/users/login/kakao", "/api/v1/users/email-auth", "/api/v1/chat/*"}) //필터 등록, 필터링 url 설정
 //@WebFilter(urlPatterns = {"/api/v1/chat/*"}) //필터 등록, 필터링 url 설정
-@WebFilter(urlPatterns = {"/api/v1/chat/*", "/api/v1/rooms"}) //필터 등록, 필터링 url 설정
+@WebFilter(urlPatterns = {"/api/v1/users/*", "/api/v1/chat/*", "/api/v1/rooms/*"}) //필터 등록, 필터링 url 설정
 public class JwtAuthenticationFilter implements Filter {
     private final JwtTokenProvider jwtTokenProvider;
     private final AuthService authService;
